Allow customizing the naming convention for log files to use timestamps.
We find UUIDs to quite awkward in the field because the files cannot be
sorted in chronological order by filename. This patch adds an optional
convention to the DDLogFileManagerDefault class which uses timestamps
instead. The existing UUID-based naming convention is retained as the
default convention for backwards compatibility.

Timestamps are formatted as RFC 3339 dates in the UTC time zone.