Expose CLIColor header publicly in cross-platform framework target
In #745, CLIColor.h was made public. However, there was not a corresponding addition to CocoaLumberjack.h to make this file publicly available. This should fix that issue.

This results in the following error when attempting to import CocoaLumberjack (when built via Carthage):
```
Umbrella header for module 'CocoaLumberjack' does not include header 'CLIColor.h'
```