Fix nullability warning when appending potentially null string
We know that the array is not empty at this point, and we know that
componentsSeparatedByString won't return NSNull values, so the fallback
will never be used. If it were used however, it wouldn't change the
newFileName, which makes it safe.

Full warning:

implicit conversion from nullable pointer 'id _Nullable' to non-nullable
pointer type 'NSString * _Nonnull'